+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 * Icon Button: Nieuw component ([#2315](https://github.com/dso-toolkit/dso-toolkit/issues/2315))
 * Tijdreis Banner: Nieuw component ([#3246](https://github.com/dso-toolkit/dso-toolkit/issues/3246))
+* History Items: Nieuw component ([#3192](https://github.com/dso-toolkit/dso-toolkit/issues/3192))
 
 ### Fixed
 * Legend Item: Styling symbols ([#3283](https://github.com/dso-toolkit/dso-toolkit/issues/3283))